thrid_point

+++ database/src/main/java/by/itacademy/dao/interfaces/VoteDao.java
@@ -1,8 +1,6 @@
 package by.itacademy.dao.interfaces;
 
-import java.math.BigDecimal;
-
 public interface VoteDao {
 
-    BigDecimal findByBookId(Long bookId);
+    Double findByBookId(Long bookId);
 }
